Move list fixups into a library
